Integration test tweaks

+++ src/test/java/org/oscarehr/common/dao/utils/SchemaUtils.java
@@ -61,7 +61,9 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang.NotImplementedException;
@@ -159,37 +161,81 @@ public class SchemaUtils
 		String osName = System.getProperty("os.name");
 		return osName.toLowerCase().contains("windows");
 	}
-	
+
+	public static Set<String> getFailedChecksums()
+		throws SQLException, InstantiationException, IllegalAccessException, ClassNotFoundException
+	{
+		Connection connection = getConnection();
+		Statement statement = connection.createStatement();
+
+		String schema = ConfigUtils.getProperty("db_schema");
+		statement.executeUpdate("use " + schema);
+
+		Set<String> errors = new HashSet<String>();
+		for (String tableName: createTableStatements.keySet())
+		{
+			// Check if there are any modified tables
+			String checksumValueSql = "CHECKSUM TABLE `" + tableName + "`";
+			ResultSet checksumResult = statement.executeQuery(checksumValueSql);
+
+			String defaultChecksum = null;
+			String checksum = null;
+			if(checksumResult.next())
+			{
+				checksum = checksumResult.getString("Checksum");
+			}
+
+			String checksumValueDefaultSql ="CHECKSUM TABLE `" + tableName + "_maventest`";
+			ResultSet checksumDefaultResult = statement.executeQuery(checksumValueDefaultSql);
+
+			if(checksumDefaultResult.next())
+			{
+				defaultChecksum = checksumDefaultResult.getString("Checksum");
+			}
+
+			if(checksum != null && !checksum.equals(defaultChecksum))
+			{
+				String errorMessage = "***** Checksums don't match for " + tableName;
+				logger.error(errorMessage);
+				errors.add(errorMessage);
+			}
+		}
+
+		return errors;
+	}
+
 	public static void restoreTable(boolean includeInitData, String... tableNames) throws SQLException, InstantiationException, IllegalAccessException, ClassNotFoundException
 	{
 		long start = System.currentTimeMillis();
-		String schema=ConfigUtils.getProperty("db_schema");
+		String schema = ConfigUtils.getProperty("db_schema");
 
-		Connection c=getConnection();
-		Statement s=c.createStatement();
+		Connection connection = getConnection();
+		Statement statement = connection.createStatement();
 		try
 		{
-			s.executeUpdate("use "+schema);
-			s.executeUpdate("set foreign_key_checks = 0");
-			for (int i = 0; i < tableNames.length; i++) {
+			statement.executeUpdate("use "+schema);
+			statement.executeUpdate("set foreign_key_checks = 0");
+			for (int i = 0; i < tableNames.length; i++)
+			{
 				String tableName = tableNames[i];
-				if (isWindows()) {
+				if (isWindows())
+				{
 					tableName = tableName.toLowerCase(); // make it case insensitive by default
 				}
 				String sql = "truncate table " + tableName;
-				s.executeUpdate(sql);
-				if(includeInitData)
+				statement.executeUpdate(sql);
+				if (includeInitData)
 				{
-					s.executeUpdate("insert into " + tableName + " select * from " + tableName + "_maventest");
+					statement.executeUpdate("insert into " + tableName + " select * from " + tableName +
+						"_maventest");
 				}
             }
-
 		}
 		finally
 		{
-			s.executeUpdate("set foreign_key_checks = 1");
-			s.close();
-			c.close();
+			statement.executeUpdate("set foreign_key_checks = 1");
+			statement.close();
+			connection.close();
 		}
 		long end = System.currentTimeMillis();
 		long secsTaken = (end-start)/1000;
@@ -212,6 +258,10 @@ public class SchemaUtils
 			s.executeUpdate("use "+schema);
 			s.executeUpdate("SET foreign_key_checks = 0");
 			for (String tableName:createTableStatements.keySet()) {
+				if("dashboard_report_view".equals(tableName))
+				{
+					continue;
+				}
 				try
 				{
 					s.executeUpdate("truncate table " + tableName);
